+++ api/src/main/java/org/allaymc/api/entity/component/EntityProjectileBaseComponent.java
@@ -11,12 +11,12 @@ public interface EntityProjectileBaseComponent extends EntityBaseComponent {
      *
      * @return the entity that shots the projectile.
      */
-    Entity getShootingEntity();
+    Entity getShooter();
 
     /**
      * Sets the entity that shots the projectile.
      *
-     * @param entity the entity that shots the projectile.
+     * @param shooter the entity that shots the projectile.
      */
-    void setShootingEntity(Entity entity);
+    void setShooter(Entity shooter);
 }